Part 4 done

+++ bloglist/controllers/blogs.js
@@ -1,6 +1,20 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
 const User = require('../models/user')
+const jwt = require('jsonwebtoken')
+const middleware = require('../utils/middleware')
+
+const getUserByToken = async (request, response) => {
+  if(request.token === undefined || request.token === null || request.token === ''){
+    return response.status(401).json({ error: 'token invalid' })
+  }
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+    const user = await User.findById(decodedToken.id)
+    return user
+}
 
   blogsRouter.get('/', async (request, response) => {
     const result = await Blog.find({})
@@ -8,10 +22,10 @@ const User = require('../models/user')
     response.json(result)
   })
 
-  blogsRouter.post('/', async (request, response) => {
+  blogsRouter.post('/', middleware.userExtractor, async (request, response) => {
+    const user = request.user
     const blog = new Blog(request.body)
 
-    blog.title == undefined || blog.url == undefined
     if(blog.likes == undefined){
       blog.likes = 0
     }
@@ -19,14 +33,11 @@ const User = require('../models/user')
       return response.status(404).end()
     }
 
-    const createdUser = await User.findOne({})
-    console.log(createdUser)
-    blog.user = createdUser.id
-
+    blog.user = user.id
     const result = await blog.save()
 
-    createdUser.blogs = createdUser.blogs.concat(result._id)
-    await createdUser.save()
+    user.blogs = user.blogs.concat(result._id)
+    await user.save()
     response.status(201).json(result)
   })
 
@@ -39,9 +50,16 @@ const User = require('../models/user')
     }
   })
 
-  blogsRouter.delete('/:id', async(request, response) => {
-    const result = await Blog.findByIdAndDelete(request.params.id)
-    response.status(204).end()
+  blogsRouter.delete('/:id', middleware.userExtractor, async(request, response) => {
+    const user = request.user
+    const result = await Blog.findById(request.params.id)
+    if(user.id.toString() === result.user.toString()){
+      await Blog.findByIdAndDelete(request.params.id)
+      response.status(204).end()
+    }
+    else{
+      return response.status(401).json({ error: 'permission denied' })
+    }
   })
 
   blogsRouter.put('/:id', async (request, response) => {
